Suppress manifest by kind
The current implementation suppresses manifests using the template key from the ConfigMap that stores the template. Although using the resource kind as the template key is recommended, it is not always guaranteed. This commit changes the approach to extract the resource kind directly from the manifest and apply the suppression accordingly.

Signed-off-by: Sharat Akhoury <sakhoury@redhat.com>