Refactor optional struct fields in ClusterInstance (#44)
This commit refactors the optional struct fields in the ClusterInstance CRD by changing them to a pointer of that struct. This allows the `omitempty` json tag to then safely omit an unspecified field from the generated CR.

The `anyFieldDefined` template function is removed as it is not needed with the refactoring introduced in this commit.

Signed-off-by: Sharat Akhoury <sakhoury@redhat.com>