Add ObservedGeneration to ClusterInstance.Status (#58)
This commit introduces an API change to ClusterInstance.Status. Specifically, the ObservedGeneration field is added. This field is updated to match the ClusterInstance.ObectMeta.Generation field each time an update/patch request is made to the ClusterInstance.

This change allows for avoiding unnecessary reconciles of the ClusterInstance CR which can be triggered when there is no change in the ClusterInstance spec fields and the controller pod is restarted (either due to an error or an upgrade).

The PR also introduces cosmetic changes by renaming some of the update/patch/find functions in conditions.go. Additionally, unit-tests are added to extend the test-coverage of some of these functions.

Signed-off-by: Sharat Akhoury <sakhoury@redhat.com>