Delete manifests owned by the ClusterInstance (#76)
This commit contains the changes to enable a user to retain resources rendered by the SiteConfig operator upon deletion of the ClusterInstance object. This is accomplished by filtering on rendered resources that contain a matching ownedBy label corresponding to the ClusterInstance object.

This commit also removes the setting of the Kubernetes OwnerReference as it only applies to Namespace-scoped resources whereas the previously added ownedBy label supports ownership of both Cluster-scoped and Namespace-scoped resources. In support of this, the ClusterDeploymentReconciler is also updated to make use of the OwnedBy label instead of OwnerReference.

Signed-off-by: Sharat Akhoury <sakhoury@redhat.com>