Enhance ClusterInstance CRD: Add NodeRef to NodeSpec
This commit extends the `NodeSpec` object in the `ClusterInstance` CRD by introducing a new field, `NodeRef`. This field enables users to reference `BareMetalHost` resources in a different namespace, improving flexibility in multi-namespace deployments.

Additionally, the default installation templates for both AI and IBI have been updated to accommodate this change.

Signed-off-by: Sharat Akhoury <sakhoury@redhat.com>